pyproject.toml
[tool.poetry]
name = "rentmate"
version = "0.1.0"
description = "Open source property management assistant"
authors = ["RentMate Contributors"]
readme = "README.md"
packages = [{ include = "llm" }]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.27"
langgraph = "^0.5.4"
langchain-community = "^0.3.27"
litellm = "^1.74.8"
langchain-litellm = "^0.2.2"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
strawberry-graphql = "^0.282.0"
pypdf = "^4.0"
python-multipart = "^0.0.22"
sentence-transformers = "^5.2.3"
chromadb = "^1.0"
pyjwt = "^2.9"
sqlalchemy = "^2.0"
langchain-text-splitters = "^0.3"
nanobot-ai = "^0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
alembic = "^1.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
